ui: Fix table use-after-drop in QueryFlamegraph (#3252)
Right now, a temporary table is created containing the slice data for
the flamegraph every time a new area selection is created (i.e. every
time the area selection is created, or an existing one is modified).
QueryFlamegraph runs long-running asynchronous operations internally,
which depend on this temporary table.

Every time a new area selection is created, the old table is dropped in
order to avoid leaking memory. However, since we have no idea what the
QueryFlamegraph is doing internally, we can drop this table while the
QueryFlamegraph is in the middle of using it, which causes issues.

This change essentially shifts ownership of the underlying source table
into the QueryFlamegraph itself, so that it can be only dropped once
it's actually no longer being used.

Fixes: https://buganizer.corp.google.com/issues/447414248