ui: group settings page by plugin (#3254)
This CL groups the settings page view by the plugin which added the
setting. Note that all core_plugins are added to the "core" section
rather than being pulled out into plugins (because to all intents and
purposes, they should be considered part of the core and it's an
implementation detail where exactly they are).

Fixes: https://github.com/google/perfetto/issues/2584