ui: customizable minimum track height (#3259)
Tracks once occupied more vertical space, which can be more comfortable
especially for users with vision and motor (pointer control)
difficulties. To accommodate embedding applications and users with
accessibility needs, add support for customization of the height of
tracks via a new setting.

Besides accessibility concerns, this also relates to the customizability
of the UI for embedding applications, such as captured in issues #2166
and #2266. For example, my application that reuses Perfetto UI sets the
minimum height of tracks using this new setting (the settings UI is not
made available to my users) to match the look and feel of the rest of
the host application.

----

A demonstration of the new setting:


https://github.com/user-attachments/assets/77f12343-3f13-441a-bb53-4a413c714fbd

---------

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>