docs: update SQL Macro usecase
In PR#3276, we found a MACRO usecase which wasn't listed before.
Therefore adding this usecase and explainations of it in the doc.

Basically, it would be ok to use MACRO for a constant if it would be
referenced massive times and it would be bad to use FUNCTION for it
which will bring massive performance overhead since SQL FUNCTION can't
do inlining.