Refactor such that DESTROYED state event marks the end of a thread
Before the DEAD state event was considered the end of a thread and
DESTROYED event was ignored. The downside of this is that in the UI
threads would end up with never ending Zombie thread state slices, which
is somewhat misleading when analyzing traces. This change makes it so
that the DESTROYED state event marks the end of a thread. The DEAD state
event is still handled and visualized, but doesn't end the thread. Now,
the DEAD event can only be followed up by a DESTROYED event. Any other
event following DEAD would result in an invalid order error.

Also, the UI was updated such that gaps in between context switch events
are visualized in the thread state track as "[Unknown]" slices. These
gaps are possible because a thread could be context switched out and we
never get an appropriate event that closes the pending slice, so it
stays "hanging".

Bug: 451701588