tp: fix parsing of protos -> args when descriptor and proto disagree
It's possible that people write protos which don't say that they have packed fields
but the protos themselves write such fields. This was always possible though
unlikely though with the new rise of editions, this becomes far more likely as
packed becomes the *default* while we consider packed *not* to be the default.

As we were looking at the descriptor to make decisions on packed/not for proto_to_args
we had cases where we were misparsing protos. Instead, rely on the encoding itself
to decide whether or not a field is packed.