tp: add support for attaching callstacks to track events (#3299)
This CL implements support for people to attach a callstack directly to
a track event.
This is particularly valuable for people converting to Perfetto who may
want to
annotate the begin/end event of a slice (or an instant) with that info.

We show the callstack in two ways:
1) The callstack is direclty shown when you click on the slice
2) We have an area selection aggregation where we aggregate all the
track event
   callstacks into a flamegraph

Fixes: https://github.com/google/perfetto/issues/1338