ui: Add additional simple slice aggregator to fix missing aggregator (#3270)
Since #2946, slice tracks must contain a `parent_id` column in order to
be considered for slice aggregation, which rules out some of the
slice-like plugin tracks that are generated from a query and don't
necessarily have a parent_id in their data source or are tracks with no
concept of a stack or depth.

This patch relaxes the parent_id constraint, allowing tracks both with
and without parent_id to be aggregated.

Fixes https://buganizer.corp.google.com/issues/448068359