ui: Preserve 'Current Selection' tab panel state (#3351)
While other tab's state is preserved when switching tabs, the 'Current
Selection' tab was left out, so switching away from it would cause it to
lose state.

This patch fixes this, allowing the current selection tab's state to be
preserved including the currently selected sub-tab in aggregations.

Fixes: https://buganizer.corp.google.com/issues/453942809

Note: This fix doesn't persist state between sub-tabs in the current
selection tab. This is a different issue, and will be fixed in a
followup.