exp: Cleanup (#3366)
Remove a lot of obsolete logic, that just stayed there since different
implementations of QueryBuilder:
- Ports rendering
- Separate styling for blocks and single nodes
- A lot of small cleanups here and there

Remove materialisation as maintaining this code was costly at this stage
of the project.

Consolidate he nodes creation logic: NodeBox is responsible for what
should be inside the node, the colours and buttons regarding one node.
NodeBlock has multiple NodeBoxes, SingleNode has one, and they are
responsible for handling the borders and special behaviours regarding
each type of graph object. NodeContainer is responsible for state
management and drag and drop logic.