sma: fix unnecessary extra calls to FlushPendingCommitDataRequests (#3355)
In the case where the SMA was >50% full, we were accidentally spamming the task running
with FlushPendingCommitDataRequests calls. There's no point in doing this as one call
to this basically does the same as n calls. However, it does cause a bunch of CPU
usage in various processes using the SDK (traced_probes, surfaceflinger) unnecessarily

Remove this excessive spamming of the task runner by only posting the task once
to the queue. Any future calls are simply ignored. We put this behind an Android flag
to incrementally roll out this change.