sma: fix not posting immediate flushes with shared memory emulation (#3374)
https://github.com/google/perfetto/pull/3355 fixes the case for normal
flush scheduling but didn't properly handle shared memory emulation.
It's very likely that shared memory emulation is hitting a *very*
similar pattern of `UpdateCommitDataRequest` being called many times in
a loop without yielding to the task runner.

Therefore, speculatively apply the same fix assuming that the same issue
is being hit in that context as well. It also has the benefit of keeping
the two codepaths very similar which is good to reduce the amount of
unique conditions we need to think about here.