tp: decouple DeobfuscationTracker from StackProfileTracker
This is important because there is a hidden assumption with
StackProfileTracker that frame names are mutable but this is not true on
perf.data files. Because of this, we were getting out of date data.

Instead, at the end of the trace, go through the frames directly. This
is a much more scalable solution