tp: add new import logs table and prevent negative durations for slices (#3357)
This CL adds a new tracker and table for storing logs which happened during import.
The idea is that instead of DLOG/ELOG, we should write to this table so that we have
a queriable way to keep track of import errors beyond just the stats table which
is very generic.

Use this new table for the first time to track negative durations in slice tracker
and gpu work period: we've seen a bunch of traces hit these and it's best to track
them properly.