intrinsic: Stack instead of heap alloc for ii struct
Since ii() maxes out at 5 tables, this data structure can be statically
allocated to 5 tables with a std::array instead of std::vector. In
theory, this is less overhead in setting up and has better cache
locality. In practice, there are non-negligible performance improvements
(data in the patch commit text).

Bug: 454944449
Signed-off-by: Samuel Wu <wusamuel@google.com>