ui: Add tracks specifically to the default workspace (#3432)
Currently all tracks that are added on trace load are added to the
current workspace. Really they should be adding them to the default
workspace, and this could cause issues if the user changes the workspace
while tracks are still being added on startup.

This patch exposes `defaultWorksapce` in the Trace interface, and
renames what is currently just called `workspace` to `currentWorkspace`
for clarity. All tracks added on startup now use `defaultWorkspace` and
all tracks added on user action e.g. via a command are added to
`currentWorkspace`.