rust_sdk: Add initial version of Rust SDK (#3330)
This SDK provides Rust bindings for Perfetto on top of the C ABI.

Designed to provide efficient, safe and easy to use bindings for the
Rust language. Includes complete documentation with examples and a
comprehensive set of unit tests.

Following components accomplish this:

- perfetto-sys crate: Serves as a low-level binding layer to the shared
C library.
- perfetto crate: Provides a safe interface and is recommended for
normal usage.
- perfetto-derive crate: Convenient proc macros for tracing function
calls and automatically capturing all input parameters.
- protozero_rust_plugin: Used to generate rust bindings for a selected
set of protos.

Like the public C headers, a complete set of rust protos are not
generated by default but are instead limited to a smaller set of protos
that can be expanded as needed. The rust macros used to produce bindings
also make it easy to include more protos or add custom protos by
projects using this SDK.

The prefetto-sys crate has two modes, "vendored" and "non-vendored". The
latter is the default and depends on an existing C shared library. The
first relies on amalgamated source for the C shared library, which is
built and automatically linked when the crate is compiled.

The perfetto crate can take advantage of branch prediction hints when
built using a toolchain that supports this. The "intrinsics" feature can
be used to enable this.

Fixes: https://github.com/google/perfetto/issues/3329