perfetto: introduce proper absl-esque hashing framework
This CL introduces a hashing framework based on absl's Hash
implementation which uses Argument Dependent Lookup (ADL) instead of
specializations of std::hash in the std:: template. This is much eaiser
and less verbose to understand and can be done in a way which is generic
to the hash function which is a huge plus.

For making this work, rewrite MurmurHashCombiner to be compatible with
this new hashing framework. Also reimplement all of MurmurHash functions
on a better base with a more intuitive heirarchy of which function
depends on what.

Overall, this makes it so much nicer to implement hashes going forward.