ui: Cache vdom for individual components of SqlTableTab. (#3423)
With the recent changes to Grid, the vdom now stores the adjusted column
widths,
so it's very counterproductive to recreate it.

Refactor SqlTableTab implementation to wrap the child views into Gate,
ensuring
that the vdom is preserved when changing views.

Also update the spelling of some variables and switch the implementation
to use TabStrip.