stdlib: Workaround broken startup data causing startup plugin crash
The startup_breakdowns module is used by the UI startup plugin.
The module assumes that we cannot have concurrent startups on the
same utid. Which should be true, except it appears there's a platform
bug breaking that, b/456092940.

Workaround the crash by pre-filtering the startups to only pick one
in any concurrent group.

Uses self-intersect to identify concurrent groups and then filters the
group to remove all but one startup.

Test: Manual
Change-Id: I91b733b4cfb2f588fe2f4152b5c44080fccfb979