ui: Add PerfettoSqlType (#3331)
Add a representation of a PerfettoSQL type in typescript and add an
ability to parse it from a raw string.

Use it to generalise generating table definition for the table viewer.

This will be used in the future to teach the table viewer about the
PerfettoSQL types and customise the behaviour based on the types.