ui: fix^2 crash when selecting track event samples
Was playing around with this and I realised that there's a small edge
case: we don't want to totally exclude the slice when it has negative
duration, we just want to treat it the same as having an instant:
this is because the start of the slice might still have an associated
callstack which we want to include.

The rest of the logic of this query does the right thing if you
add a max here.