tp: improve error handling of clock synchronization and use import logs (#3468)
This CL improves how clock synchronization erorrs works: instead of just
having a single "clock sync failed" error, break it down into multiple
smaller errors. Also use the new error logging system of trace processor
to track those errors with much more fine grained details plumbed to the
import log tables. This makes it easier to understand what is going on
without needing to use traceconv text.

To be able to do this, the ClockTracker API has been refactored to no
longer
return a hard error but a soft error via std::optional (as
realistically, it should
never be the case that a single clock sync causes a whole trace to fail,
especially in the presence of merged traces). We move the logging inside
clock tracker as we can pass fine grained information which it would be
cumbersome to make the caller deal with.