tp: migrate track event to new import logging system (#3469)
track event is the Perfetto protos which are written the *most* in
Perfetto. For this reason, it's among the most high priority thing for
people to debug if there are subtle breakages.

Fix this by making the stats much more finegrained and adding them to
the import logs table which will soon be displayed in the UI. Remove
the old coarse grained errors completely as they have been subsumed.