ui: rewrite "Info and Stats" as new "Overview" page (#3455)
This CL completely overhauls the anicent "Info and Stats" page
of the UI to be much more modern and useful by sectionizing into
tabs and highlighing the most useful parts in a landing page.

Moreover, this also exposes the new import logs table of TP
in a new UI tab which was the main motivation for this CL in
the first place.