ui: fix selecting columns for pivots
Before this patch, pivot table could only add pivots and aggregations
from the original columns of the table. With the ability to cast and
transform columns, this is not sufficient, so fix the selection menu
to also include the columns that have been added to the table by the user.