Add function to parse AndroidReportConfig from the trace file (#3507)
On Android, when uploading traces after reboot, we need to know where to
upload each trace. This information is a part of a `TraceConfig`, so we
parse it back from the trace file.

Perfetto SDK users can emit the package that looks like a `TraceConfig`,
so we check `trusted_uid` of the package to make sure the 
`AndroidReportConfig` is legit.