Return the entire `TraceConfig` when parsing the trace file (#3527)
When uploading traces, we actually also need a trace `uuid`, not only
the `AndroidReportConfig`. We now return the full `TraceConfig`
to extract all the necessary information from it later.

Also change function signature: we should always check if the trace is
empty before `mmap`-ing it.

This is a follow-up to bdbc36a798c7333fbb98a9ae411ff86328f94532.