shared_lib: Set shmem_page_size_hint_kb to 4.
Matches the PerfettoProducerBackendInitArgsSetShmemSizeHintKb
documentation where callers are told to that a multiple of 4KB
must be used.

No actual change in behavior with a release build of the tracing
service as a 0 value will be adjusted to become 4KB but it removes
an "Invalid configured SMB sizes" message when using a debug build
of the tracing service.