Presubmit: improve merge-base detection (#3586)
Today run-presubmit uses origin/main as a merge-base to figure
out the list of changed files. That creates problems if you git fetch
and your branch goes behind newer changes as all the files changed
by the new changes are also considered.
Instead use git merge-base to figure out the common
ancestor and diff against that.

This does NOT affect CI, as the CI passes an explicit --merge-base