ui: Simplify AppImpl and TraceImpl (#3617)
AppImpl and TraceImpl have become bloated over time, and every new
addition to one of the usually involves adding boiiler plate to 4
different classes. It's a mess and this patch cleans it up.

- Remove plugin specific features (pluginId, plugin-specific openerArgs
and routerArgs) from polluting the App and Trace interfaces, these are
injected into the plugin separately.
- Remove a large amount of plugin related boilerplate code from AppImpl
and TraceImpl.
- Move all plugin-oriented proxying (e.g. to fill in plugin ids when
registering entities) to the plugin manager where it belongs.
- Move startup command execution into load_trace.ts rather than have it
injected into the command manager's proxy (for some unknown reason!).
- Fix tagging query logs when using an EngineProxy.