Use process table to identify machine ids in Android logs
When ftrace isn't set as a data source during a multi-VM tracing
session, the CPU table won't be populated. This means that if Android
logs were indeed set as a data source, then the AndroidLog panel won't
show the machine id column and filter button, even though there are logs
from multiple machines, because it uses the CPU table to identify the
machines involved in the table.

Since every log has a matching UPID then we can iterate through the
process table instead and capture the machine ids from there, assuming
that we will have way less processes than logs in the trace.

Bug: 459819089