Use machine table to identify machine ids in Android logs (#3624)
When ftrace isn't set as a data source during a multi-VM tracing
session, the CPU table won't be populated. This means that if Android
logs were indeed set as a data source, then the AndroidLog panel won't
show the machine id column and filter button, even though there are logs
from multiple machines, because it uses the CPU table to identify the
machines involved in the table.

Instead we use the machine table which stores all the machines involved
in the multi-VM trace.

Bug: 459819089