Cts reporter tests: cleanup test data before and after test run.
When running test locally with `atest CtsPerfettoReporterTestCases` tradefed uses the "Incremental Setup" feature to speedup test execution. With this feature it doesn't re-install the test APKs if they doesn't change. In our tests, even though the APK doesn't change,
it adds more files to it internal storage after each test run.
In this CL we add a command to cleanup the internal data before and after each test run
even when the APK is not re-installed.