Remove another superfluous use of app.trace
Mod+P is ued to find tracks. Tracks are purely a trace concept. This
command was being registered in onActivate using app.trace to find the
trace. This patch moves it to onTraceLoad and captures the trace in the
command callback.

A manual block of the print dialog has been put in place to avoid
confusion before a trace is loaded.

Change-Id: I7aac44de281dd5bb9214056506ebc982b6538ac5