ui.perfetto.dev: support properly gzip transparent encoding (#3689)
This PR fixes our server-side GAE (AppEngine) to properly
support transparent gzip compression. This is how it works:

**Storing/Serving in GCS**:
When we upload files to gcs via the build_all_channels.py file
need to pass `-z json,html` to gsutil cp. This makes it so that the
file is stored in GCS with the`Content-Encoding: gzip` metadata.
This makes it so that GCS, when serving the file:
 - Transparently decompresses it, and serves the cleartext, if the
   HTTP client did NOT specify any `Accept-Encoding` header.
 - Passes through the compressed stream if the HTTP client says
   `Accept-Encoding: gzip`
IMPORTANT NOTE: we MUST NOT set the Cache-Control headers at gsutil
time.
 Setting those seems to mess with the transparent serving mechanism
 described above. These must be left to the default.
 The cache behaviour is instead set by GAE.

**GAE changes**:
There was a bug in the previous attempt I made, due to the fact
that the python 'requests' module supports gzip encoding, but does
so transparent. This would end up in a situation where we propagate
in output the "Content-Encoding: gzip" header received from GCS, but
then we'd return the full uncompressed content as decoded by the
request object.
This new change instead uses raw.read() which accesses the underlying
raw content.
Also move the cache-control header here.