c_sdk: fix C++ code in C-clean headers
There were some small issues in the C headers which i found as I was
testing this code for https://github.com/google/perfetto/pull/3701.

Basically it seems like we were not correctly doing struct X and we were
accidentally using nullptr which is fine on C++ but strict C compilers
fail.