tp: fix power rail timestamp_ms clock conversion
per protobuf spec the timestamp_ms in energy data is supposed to be
in CLOCK_BOOTTIME, however the code currently inserts it as is
assuming it is the trace timestamp.

CLOCK_BOOTTIME often is the trace time, so this usually works fine, but
if a different domain is used for the primary trace clock, then this can
fail.

Update logic to convert to trace time if timestamp_ms is set.

Also update one packet in existing diff test to ensure it gets converted
correctly, so that test now verifies both code paths:
(clock sync with packet ts and clock sync with energydata ts)