ui: make the caller deal with query flamegraph state
We were jumping through *lots* of hoops to juggle flamegraph state and
it was a lot harder than it needed to be. A lot of that came from the
fact that we were trying to manage the state updates of the
QueryFlamegraph instead of letting the caller (who might embed things in
a struct or use a store) deal with it.