ui: preserve flamegraph state across different selections (#3706)
This CL preserves the state of flamegraph across different selections of
the same "type" of flamegraph (i.e. preserving across different
callstack selections or different slice selections but *not* between
callstack and slice selections).

Also while I'm here, make some improvements to how slice flamegraph
handling works to reduce flicker (we were keeping around the old
flamegraph for a bit and then flicker to loading and then new
flamegraph).

In a followup change, we'll also persist this state as well in
permalinks