ui: add permalink serialization for all flamegraph users (#3710)
This CL makes all flamegraph users serialize their state. We store this
state in the store at the plugin level and pass it down through the
layers.

There's a small catch: details panel *already* serialized their state
into the selection object. We have to retain this to avoid breaking
backcompat for everyone. There's some awkward code shuffling but nothing
which is too egregious.