ui: move metrics to render function also and fix crashes
Since recent past refactors, we were accidetally crashing if render()
was called before load() as the assertExists would fail. But the only
reason QueryFlamegraph was null in the first place was because of not
having state ready for it.

ALl of this can be avoided if we just pass the metrics to the render
function on every cycle. So just do that. This requires rejigging a
bunch of code but I think this makes things *much* easier to follow now.

Fixes: b/461794263