ui: use crossorigin only on fonts in link rel="preload" (#3738)
Follow up to #3692. I misread the docs and didn't realize
that the crossorigin attribute is needed only for fonts
(Which are subject to CORS) but NOT css.
This ended up invalidating the preload for the CSS, showing
a devtools warning that Lalit spotted.
Also turns out that jsut setting "crossorigin" is enough.
See
https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/link#anonymous
```
Note: font and fetch preloading requires the crossorigin attribute to be set; see CORS-enabled fetches below.
```

Bug: b/458318781