tp: refactor trace_processor_shell to extract library target (#3664)
In Google3, we want a version of shell which is tightly integrated with
the rest of Google infra (e.g. the SQL modules, the internal filesystems
etc). But we don't want to maintain a separate shell.

The best way around this is to extract a thin wrapper interface for
shell which allows restricted customization points of shell without
needing to duplicate lots of code