multi-machine tracing: stop matching DSes across machines by default (#3714)
This is a breaking change to the default data source matching behaviour
across remote producers (connected over traced_relay). Before, we
matched across machines, now we match only the host unless the config
sets the new "trace_all_machines" setting to true. In either case,
DataSource.machine_name_filter can be used to set an explicit set of
machines to match.

The current default is surprising because perfetto has historically been
single-machine, so existing users and configs do not expect to suddenly
match more data if a VM with traced_relay is added to their device.

Note that I am keeping the "builtin producer" behaviour of emitting
system_info *and* remote_clock_sync packets for connected traced_relays,
so the trace will reflect that there's a secondary machine even if none
of its producers matched the trace config.

Bug: 460163124
Tested: `atest CtsPerfettoTestCases` on a frankel with a microdroid VM
        running traced_relay + traced_probes