ui: add api for snapping to items in tracks and implement for slices
This CL implements a new "snapping" idea to the track API: this is a new
fully optional API for tracks to implement which allows the user when
doing area selections to "snap" to the boundary (or even some internal
timestamp) of some item in the track.

The main motivating reason for this is slice boundaries: people often
want to measure the "inter slice" duration or "the duration from start
of slice A to end of slice B": in these situations if they wanted more
precise measurements, they'd have to zoom in on that region.

Instead, with "snapping" it goes to the nearest slice when the user's
cursor is "close" to the slice.

Snapping occurs in the following situations:
* end of an area selection is close to a slice
* start or end drag handles are moved close a slice
* importantly it does *not* work for area selection start: I personally
think snapping here would feel weird to the user.

As for how it is visually implemented:
- we draw a vertical "dashed line" for the full length of the UI
  whenever we snap to a slice.
- hypothetically we could also draw something on the slice itsemf but in
  practice I haven't found that to be necessary: maybe something to add
  in the future if it becomes necessary

Fixes: https://github.com/google/perfetto/issues/154