base: MacOS: align behavior of UnixSocket::SetTx/RxTimeout (#3744)
On MacOS, calling setsockopt() on a VALID server socket when
the peer socket has disconnected causes an EINVAL.
This is: (1) annoying, because the server cannot possibly be
checking for disconnection on every instruction; (2) different than
what Linux does; (3) Feels wrong: just because the other end has
disconnected, I don't see why I shouldn't be able to change the
timeout of my own end, which is still valid.
This PR aligns the behavior of SetTx/RxTimeout on mac with what
happens on Linux.

Bug: b/462117635