Fix protozero_descriptor_diff deps (#3768)
The `deps` on the binary need to match the `binary_label`.

The mismatch appears to be causing Chromium's `win11-arm64-blink-rel`
trybot to fail. This is a x86_64 host building for arm64. The build
rules currently state to use the host binary for the action but then
only depend on the target binary being built.