perfetto: add support for amalgamating C SDK (#3701)
This CL adds support for generating a single header + cc file for the
C SDK. This requires making sure that only C code is visible in the
headers: note that the implementation is still a C++ file because it
still relies on the C++ SDK under the hood: only the API is C clean.

I've tested this and made sure it compiles with an example program.

Fixes: https://github.com/google/perfetto/pull/3701