GN: split `perfetto_unittests` binary in standalone builds (#3748)
We want to speed up the feedback loop when working on different
parts of the Perfetto code base. It means that we don't want to
build e.g., the TP if we work
on code in `base/`.
Before this CL, all our unittests were compiled into a
single binary, so we had to rebuild a lot of code for each
small change.

Even though we have an `enable_perfetto_trace_processor`
GN arg, many targets unrelated to TP won't build when it's
disabled.

This CL adds separate test binaries for different parts of the perfetto
project:
```
//:perfetto_base_unittests
//:perfetto_misc_unittests
//:perfetto_profiling_unittests
//:perfetto_trace_processor_unittests
//:perfetto_tracing_unittests
```

This CL only affects the standalone builds; it is a no-op for both
embedders
and generators. 

Also remove old (and we believe no longer used) `is_fuchsia` conditional
dep from the unittests targets.

Bug: b/447162132