ui: Resolve JavaScript stacktraces in the browser (#3731)
Decode JavaScript stacktraces in the browser to TypeScript sources in
the crash dialog and when reporting crash reports to buganizer.

Before:
```
 - Object.callback (frontend_bundle.js:95363:24)
 - CommandManagerImpl.runCommand (frontend_bundle.js:216751:27)
 - onSubmit (frontend_bundle.js:338941:27)
 - HTMLInputElement.onkeydown (frontend_bundle.js:339162:34)
 - EventDict.handleEvent (frontend_bundle.js:10769:56)
```

After:
```
 - Object.callback (src/plugins/dev.perfetto.Chaos/index.ts:28:14)
 - CommandManagerImpl.runCommand (src/core/command_manager.ts:98:20)
 - onSubmit (src/frontend/omnibox.ts:179:17)
 - HTMLInputElement.onkeydown (src/frontend/omnibox.ts:523:20)
 - EventDict.handleEvent@node_modules/.pnpm/mithril (2.2.2/node_modules/mithril/render/render.js:890:54)
```

This consists of two parts:

1. Build system

A new rollup plugin injects a snippet of code at the end of each bundle
which injects a stripped down source map for that bundle into a registry
called `__SOURCEMAPS` in the global namespace.

```js
self.__SOURCEMAPS = (self.__SOURCEMAPS || {})['<bundle-file-name>']=<source-map-content>
```

Note: Bundles loaded into workers (e.g. engine and traceconv) have a
different `self` object to the frontend bundle, so this registry is not
for them. The reason this is a registry rather than a single variable is
if we ever need multiple bundles in the main UI thread in the future I
want to avoid this biting us.

A new dev dependency has been added on `source-map` which is used to
strip down the source map before embedded it.

2. UI

The UI can now use this sourcemap registry to look up source mappings
for each line in a stack trace. This translated stack is then displayed
in the crash dialog and used for analytics.

**How to Test:**
1. **Trigger a crash:** Run the command 'Add debug slice track' and
enter an invalid query.
2. **Check the crash dialog**: It should show typescript filenames and
line numbers.