ui: Remove "supporting columns" from SqlTable (#3591)
Remove the concept of `supporting columns` from the SqlTable. It
originally served
to allow richer visualisation of the data in the table (e.g. to render
arg value
correctly we need to know its value, as well as its type). Supporting
columns allowed
for multiple sql values to be fetched when rendering a given
TableColumn.

A much better approach is to fetch a single complex value (in protobuf
or json) and
render it. To support this, TableColumn.display is introduced:
TableColumn.column is
still considered the "canonical" underlying value (used for aggregation,
casting, etc),
while, if set, `TableColumn.display` is passed to `renderCell`.

This patch also converts the only existing use case for "supporting
columns": a new
__intrinsic_serialise_arg(arg_set_id, key) SQL function is introduced,
which returns
a ArgValue proto.

Also clean up some complexity in query generation which is no longer
needed.