ui: introduce a new set of presets and signifcantly improve record page (#3780)
This CL adds the idea of "presets" to the recording page for Android and
Linux. This basically is "preset saved configs": configs people can
start recording traces on the UI without needing to go and manually
configure knobs.

While doing this, I also realised we should really merge the cmdline
instructions, saved configs and record pages into one page which
efficiently lays out all the information: there was no need for that to
spread across three pages.

So implemented a nice merged page using all the great card/grid widgets
similar in style to the overview page. The idea now is that you
literally don't need to leave the overview page to record a trace:
everything can be done from that one page.